Fix: standardize competitionId types from Int to Long
Bug Risks (HIGH): Results module used Int for competitionId while other modules used Long, risking silent overflow for IDs exceeding Int.MAX_VALUE

- Added tests in CompetitionIdTypeConsistencyTest
- Changed competitionId to Long in ResultsRepository, ResultsDao, ResultsRemoteDataSource, Screen routes, AppNavHost navigation args, ResultsViewModel, ShooterResultViewModel, and all callers

+++ app/src/main/java/se/kjellstrand/webshooter/data/results/remote/ResultsRemoteDataSource.kt
@@ -8,6 +8,6 @@ interface ResultsRemoteDataSource {
 
     @GET("/api/v4.1.9/competitions/{id}/results")
     suspend fun getResults(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
     ): ResultsResponse
 }
